fix: remove unused exports for nextjs typed routes
